+++ codex/librarian/crond.py
@@ -36,9 +36,9 @@ class Crond(Thread):
         """Intialize this thread with the worker."""
         super().__init__(name="crond", daemon=True)
 
-    def shutdown(self):
+    def stop(self):
         """Stop the cron thread."""
         self.run_thread = False
         with self.COND:
             self.COND.notify()
-        self.join(self.SHUTDOWN_TIMEOUT)
+        # self.join(self.SHUTDOWN_TIMEOUT)
